🐛 FIX BUILD: Add missing exports and fix invalid route export
Fixes:
1. Added checkContentModeration function to lib/moderation.ts
2. Moved verifyWebhookSecret to lib/webhook-verify.ts (Next.js routes can only export HTTP handlers)
3. Merged REPORT_REASONS fix from main
